move scriptInputFiles() to Configuration+CommandLine.swift
